requirements.txt
torch==2.4.0
torchvision==0.19.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
Pillow==10.2.0
tk==0.1.0
tqdm==4.66.2
psutil==5.9.8
imbalanced-learn==0.12.0    
efficientnet_pytorch==0.7.1 